Enhance dbcreate/dbcheck to work with subdirectories holding .gld/.dat files (#250)
* Below is the primary commit

dbcreate.csh currently only creates gld and db files in the local and remote sides of the
test output directory. It does not let the user create the gld/db in a subdirectory on the
local/remote side of the test. That need is now implemented in dbcreate and dbcheck.

The syntax currently to create a mumps.gld and mumps.dat/a.dat
	dbcreate.csh mumps 2

The new syntax to create a mumps.gld and mumps.dat/a.dat under the dir1 directory is
	dbcreate.csh dir1/mumps 2

i.e. instead of "mumps", one needs to specify "dir1/mumps".

Similarly, the syntax currently to check the mumps.dat/a.dat db is
	dbcheck.csh

The new syntax to dbcheck mumps.dat/a.dat under the dir1 subdirectory is
	dbcheck.csh dir1/mumps

The new syntax causes both dbcreate and dbcheck to go into dir1 and see if a mumps.gld file
exists (because dir1/mumps was specified) and if so it will use that as gtmgbldir/ydb_gbldir
and do the mupip create or mupip integ -reg "*" respectively.

* Below are later commits which addressed test failures.

1) Do not change dbcheck output format as it causes test failures

2) Existing <dbcheck.csh mumps> usages should continue to do MUPIP INTEG -FILE mumps.dat (not MUPIP INTEG -REG "*") or else reference files need to be fixed so use new scheme only if a subdirectory has been specified in dbcheck

3) Fix v43001b/D9C05002121 subtest failure (where absolute path name of gld was specified in init_gld.csh)