Make r120/gctest subtest failure output more user-friendly
The r120/gctest fails rarely due to some other system load.
If a test fails, the failure diff currently is not user friendly.
This commit makes it user friendly.

Example new output is the following in case the test fails

TEST-E-FAIL : cpu(3)=0.610 is 3.17x times cpu(2)=0.192 but max allowed is 3x
TEST-E-FAIL : cpu(4)=1.989 is 3.26x times cpu(3)=0.610 but max allowed is 3x

Previously there used to be a dump of all M local variables and one has to read
through gctest.m to figure out why the test failed.