[#295] New v63003/gtm8842 subtest (Tests GTM-8842 in V63003 and YottaDB/YottaDB#295)
Tests the following release notes:

When TRIGGER_MOD is restricted, attempting to ZBREAK a trigger results in a RESTRICTEDOP error, and both ZBREAK and ZSTEP actions are ignored while executing code within a trigger. Previously a TRIGGER_MOD restriction did not imply these other restrictions. (GTM-8842)

When TRIGGER_MOD is restricted, ZSTEP actions are completely ignored while executing code within a trigger. GT.M-8842 mentioned this as fixed in GT.M V6.3-003 but the fix seems to be incomplete in that the ZSTEP action was executed right after the trigger finished (YottaDB/YottaDB#295)