New r122/ydb210 subtest (tests YottaDB/YottaDB#210 in r1.22) (#244)
This tests the following release note (fixed in r1.22)

Replication Source Server startup errors always appear in the Source Server log file. In previous versions of YottaDB, if the Source Server was started from the terminal, and the Source Server startup command failed or the Source Server successfully started but later failed (e.g. with errors like NULLCOLLDIFF, REPLOFFJNLON, REPLINSTNOHIST, SECNOTSUPPLEMENTARY etc.) this error message did not show up in the Source Server log file or on the terminal where it was started (assuming that terminal still exists at the time of the error). The release notes of GT.M V6.3-004 mention that this issue has been fixed by GTM-8576 but we found the issue exists even in GT.M V6.3-004. (#210)

In addition to the r122/ydb210 subtest changes, the test framework had two changes.

1) Ensure stdin is not redirected to /dev/null in test framework scripts as that would mean even if stdin was a terminal in a caller script (e.g. r122/ydb210 subtest), the test framework would end up starting the source server with stdin pointing to /dev/null (and not to a terminal) and hence not exercise the codepath which had the issue described in YottaDB/YottaDB#210.

2) Ensure stdout/stderr is terminal before starting source server in case stdin is a terminal (src_srvr_stdin_is_terminal env var is set by caller subtest) or else the buggy #210 codepath is not being tested