Fix occasional r122/ydb210 subtest failure (#262)
We saw an in-house failure because we started the receiver in RCVR.csh but a checkhealth done immediately afterwards (in the same script) failed because it noticed the receiver server had already errored out with a REPLINSTNOHIST error (which is expected in the test flow). The fix is to set the gtm_test_repl_skiprcvrchkhlth env var to skip the checkhealth just for this receiver startup.