Enhance test framework to use POSIX plugin even in UTF8 mode (#268)
Note though that there are issues still with the POSIX plugin Makefile which will be fixed separately.

Currently the POSIX plugin Makefile builds plugin/o/_POSIX.so and plugin/o/utf8/_POSIX.so
But the latter contains object files corresponding to M mode (and not UTF-8 mode).
This is incorrect and needs to be fixed in the POSIX plugin. Until that is done, the
test system framework change to enable POSIX plugin usage cannot be used in -unicode mode.