[#297] New r124/ydb297 subtest (tests YottaDB/YottaDB#297)
Tests the following release note:

LOCK commands work correctly in rare cases when there are more than 31 subscripts at a given lock name subscript level (across all processes concurrently using a particular database file lock space) that hash to the same value. A $$^%PEEKBYNAME("sgmnt_data.lock_hash_bucket_full_cntr",REGION) now returns the # of times such a rare event was seen (and internally handled correctly) in the region named REGION. Previously, in this situation, it was possible for LOCK commands to spin-loop and incorrectly seize ownership of the lock resource from a concurrently holding process. This issue was introduced in GT.M V6.3-003 as part of GTM-8680 and so exists in YottaDB r1.20 and r1.22.