[#293] New r124/ydb293 subtest (tests YottaDB/YottaDB#293 in r1.24)
Tests the following release note:

The Update Process operates correctly when a trigger issues a SET $ZGBLDIR to switch the current global directory (without doing a NEW $ZGBLDIR). Previously it used to switch the global directory correctly the first time the trigger was invoked but later invocations of the trigger did not switch the global directory resulting in any updates in the trigger code unintentionally happening in the primary global directory with which the update process was started. A workaround in prior versions was to do a NEW $ZGBLDIR before the SET $ZGBLDIR in the trigger code.