Fix suppl_inst_B/supplementary_err test failure due to recent _POSIX.so addition to ydb_routines/gtmroutines
When switching to an older version, a script com/set_active_version.csh was being invoked which
stripped off plugin related obj dirs. That was not updated to strip off _POSIX.so element in
the list (which was added by a recent commit to set_gtmroutines.csh). This caused a dbcreate done
in the suppl_inst_B/supplementary_err test after switching to an old 32-bit build of GT.M
(e.g. V53002) to issue a DLLNOOPEN error (type ELFCLASS64) due to the presence of the 64-bit
_POSIX.so in the gtmroutines list of a 32-bit GT.M build. Now fixed to strip off _POSIX.so
just like it already stripped off other plugin components in the list.