Fix test framework so it records test status (success or failure) in D_ALL
As part of a prior commit, pre-existing logic in com/submit_test.csh was removed
as similar logic was added to com/submit.awk. But the latter only handled the case
when the test was killed by the gtm_test_watchdog.csh script. The most common case
of a test passing or failing is handled in com/submit_test.csh and should stay as
is (since PASSED or FAILED status is known only to com/submit_test.csh). So restore
the code removed in com/submit_test.csh.