Fix occasional v60000/gtm7231 subtest failure due to lack of white-box test setup during MUPIP RUNDOWN
The v60000/gtm7231 subtest failed once (in a lot of runs)with an assert failure in MUPIP RUNDOWN.
The assert failed in WCS_FLU.C because we found no dirty cache-record in the active queue but the
dirty queue cnt (cnl->wcs_active_lvl) is non-zero. This is expected since this test does kill -9s
and depending on when the kill9 happens, we could end up with this scenario.

The test already sets a white-box variable indicating this is a crash case. But it unsets this right
after the MUPIP journal extract. Whereas this white-box setup is needed even until later when we
run the MUPIP rundown. The "unsetenv gtm_white_box_test_case_enable" is now moved to the end of
the script. That should ensure the white-box setup is there all through.