Fix occasional msreplic_H/max_connections subtest failure due to non-zero value of CTL JnlPoolWrite Sleep
We had a test failure with the below diff.

36,37c36,37
< CTL JnlPoolWrite Sleep       Cntr                        0 [0x00000000]
< CTL JnlPoolWrite Sleep       Seqno                       0 [0x00000000]
---
> CTL JnlPoolWrite Sleep       Cntr                        1 [0x00000000]
> CTL JnlPoolWrite Sleep       Seqno                    1246 [0x00000000]

These counters can be non-zero in case of system load and so the test is now fixed to
allow for an arbitrary non-zero value of the "CTL JnlPoolWrite Sleep" line items
just like it already does for "CTL ReplPhs2Clnup IsPrcAlv".