New v63004/gtm8906 subtest (tests GTM-8906 in V6.3-004) (#282)
Tests the following release note:

MUPIP JOURNAL -ROLLBACK and -RECOVER handle large amounts of journal data; while there are
still limits, previously when they attempted to handle over around half a billion updates they
inappropriately failed with a GTM-F-MEMORY error. (GTM-8906)

Although the release note mentions half a billion updates, we found that the # of updates
needed to trigger the GTM-F-MEMORY in V63003 is around 55 million. So the test is structured
around this smaller number as that reduces the time taken to run the test as well.