Fix occasional v63003/gtm8844 failure (#285)
We saw a failure in a section of the test where two halt or zhalt commands are sent in succession and
the test expects the second halt/zhalt to cause the process to exit as long as they are issued within
0.5 seconds of each other. But on some slow systems it is possible the second halt/zhalt happens
after 0.5 seconds. And that would cause a test failure. To address that, the specific section of the test
is retried for a max of 3 times until we see a desirable output. This is expected to almost eliminate
failures of this kind.