Set ydb_repl_filter_timeout env var only if it was not set coming in (helps investigate occasional FILTERTIMEDOUT failures)
We keep seeing FILTERTIMEDOUT errors in the source server occasionally in test runs of YottaDB
on the ARM platform. This is inspite of the test running with an env var set to a huge timeout
of 16 minutes. And we are not sure if this is a code issue in the source server which would
otherwise hang indefinitely (but is erroring out due to the 16 minute timeout). So want to run
tests with the env var set to a huge timeout (i.e. days) so a test hang would be signaled by
the test framework and that can be investigated while the source server process is alive.
The only issue with that approach is that the test framework currently resets this env var
unconditionally at test startup even though it was set coming in. That is fixed now.