Fix occasional unicode_io/unicode_fifo subtest failure on slow systems (#288)
In a test failure in-house, Line 38 below printed a FIFO reader couldn't start message
causing the failure. We started a job in line 36 and then wait 60 seconds for it to
execute 4 lines (starting from line 103) and set ^A=1 in line 107. But it did not.
Not sure exactly how long it took to reach line 107 so the timeout is now bumped to
15 minutes (instead of 1 minute) assuming that should be more than good enough.

While at this, also fixed those places to quit if ever the (now longer) wait times out.

unicode_io/inref/unicodefifo.m
--------------------------------
   35    write jobstr,!
   36    job @jobstr
   37    for i=1:1:60 quit:^A  hang 1
   38    if 0=^A write "FIFO reader couldn't start, exiting",! quit
    .
    .
  102 unicodefifo2(file,encoding) ;
  103    set $ZTRAP="do error2^unicodefifo"
  104    write "$J:",$J,!
  105    write "file=",file,!
  106    write "encoding=",encoding,!
  107    set ^A=1