New v63002/gtm8281 subtest (Tests GTM-8281 in V63002)
Tests the following release note:

GT.M source lines, XECUTE strings and Direct Mode input all accept up to 8192 byte values. In the case of source, the length includes the terminator(s). In the case of Direct Mode, the WIDTH and/or WRAP deviceparameter characteristics may cause a lower limit. Previously the limit for source lines was 8191, and 1024 or less for the other two. (GTM-8281)(GTM-4851)