New v63004/gtm7483 subtest (tests GTM-7483 in V63004)  (#294)
Tests the following release note:

MUPIP INTEG issues a DBKEYMX error in case of a long key name stored in the DT (Directory Tree); previously, it issued a INVSPECREC with no context. This issue was only observed in the GT.M development environment, and was never reported by a user. (GTM-7483)