New v63002/gtm8760 subtest (Tests GTM-8760 in V63002)
Tests the following release note:

GT.M now properly handles environment variables whose contents is over 32K in size. GT.M reports an error if the content is larger than can fit in a local or global. Previously, environment variables whose contents was over 32K in size resulted in a SIG-11. This issue was only observed in the GT.M development environment,and was never reported by a user. (GTM-8760)