Rework the v63004/gtm8680 subtest to avoid occasional test failures
This subtest failed once in a while on all boxes both in pro and dbg.
The # of lock operations done in a second did not monotonically increase with time.
It fluctuated back and forth causing test failures because the max and min across
all the measured seconds were very different. Therefore the testing algorithm is now
reworked to wait for an iteration where we see the maximum # of lock operations happening.
Once that happens, we measure all following iterations and get the minimum # of lock
operations across all those iterations. And signal a test failure if that minimum is less
than half of the maximum #. This avoids all the occasional test failures AND lets us run
the test with pro and dbg builds AND reliably shows a failure when run with V63003 and older.