Disable v63004/gtm8916 subtest (hopefully temporarily) on systems with glibc > 2.25
Systems with glibc version > 2.25 cause the v63004/gtm8916 subtest to hang in the receiver server
in "pthread_cond_signal". The current suspicion is that it is a glibc issue with handling kill -9
of processes using pthread_mutex_*() and pthread_cond_*() calls and that it will be fixed soon
in a newer version of glibc. Temporarily disable this subtest on such systems.