Fixed occasional v63002/gtm8644 test failure (#313)
We occasionally saw test failures with the following diff.

3,5c3,5 
< \_ tcsh -f /usr/library/gtm_test/T999/v63002/u_inref/gtm8644.csh 
< \_ /usr/library/V999_R122/pro/mumps -run psforestfn^gtm8644 
< \_ /usr/local/bin/tcsh -c ps --forest -f 
--- 
> | \_ /usr/library/V999_R122/pro/mumps -run psforestfn^gtm8644 
> | \_ /usr/local/bin/tcsh -c ps --forest 
> | \_ ps --forest -f 

This is because the ps --forest output (from processtree.out which was added in a recent prior commit) had an unrelated (to the gtm8644.csh process tree) line at the
end once in a while which caused the "tail -4 | head -3" to not always give us the desired 3 lines. The
logic has therefore been changed to check for the line containing gtm8644.csh and display the 3 lines 
from then on (including the gtm8644.csh line). This works well in all cases.