Fix rare v60001/gtm7461 subtest hang due to a test timing issue (#312)
In an in-house test run, we had a test hang. This was because the instance got frozen
(by the backgrounded "mupip journal -rollback -backward" process) due to the JNLFILEOPNERR
error AFTER the foreground test script had unfrozen the instance resulting in a later call to
mupip rundown hanging due to the instance freeze active at that point.

The test flow has now been fixed so we wait for the backgrounded rollback to finish before
getting the syslog messages. We now confirm the REPLINSTFROZEN and JNLFILEOPNERR messages are seen and include the actual message in the reference file.