New v63002/gtm8718 subtest (Tests GTM-8718 in V63002) (#314)
Tests the following release note:

Attempting to set an invalid string into $ZROUTINES leaves the previous value of $ZROUTINES as it was. Previously, this could cause issues if auto-relink was in use, resulting in a segmentation violation (SIG-11). This issue was only observed in the GT.M development environment, and was never reported by a user. (GTM-8718)