Fix com/getoper.csh script regression introduced in prior commit (#315)
As part of a prior commit (SHA id d5337e953cf35467f438bd19a8895d17ed6bd318), an enhancement
was done to better handle the case where $2 (after_time) is specified as "". In this scenario,
previous getoper.csh usages would not gather all syslog messages in transit in some cases.
But that introduced a regression in case the getoper.csh caller had also specified a $5 that
is not "". In that case, the intended use is for getoper.csh to wait for $5 to show up in the
syslog and then return whereas the enhanced getoper.csh was waiting for a random string
(generated because $2 was "") instead of $5 and so the getoper.csh call would return with the
resulting messages placed in the file specified (by $3) but would not issue a failure status
even if $5 was not found in the syslog. This actually happened in a recent test failure where
the v60001/gtm7461 subtest did the following getoper.csh call.

$gtm_tst/com/getoper.csh "$syslog_before1" "" syslog1.txt "" "REPLINSTFROZEN"

Notice $2 is "" and $5 is "REPLINSTFROZEN". In the test failure case, "REPLINSTFROZEN" was not
seen in syslog1.txt but still the getoper.csh call succeeded (the test failed due to other
reasons so we could find this getoper.csh issue) which is not correct. This is now fixed
by checking if $5 is not "" and if so (where $2 is "" and $5 is non-""), getoper.csh searches
for $5 instead of the random generated string and returns a failure if $5 is not found in the
syslog within the specified/default timeout.