Fix occasional v63001/seq_format test failure due to different sorted order of EOF journal records (#319)
We had a test failure with the following diff.

*** 205,210 ****
  09\#TIMESTAMP#\13\#PID#\0\#TOKENSEQ23#\0\#STRMSEQNO#\3\
  09\#TIMESTAMP#\16\#PID#\0\#TOKENSEQ23#\0\#STRMSEQNO#\3\
  03\#TIMESTAMP#\14\#PID#\0\24
- 03\#TIMESTAMP#\14\#PID#\0\24
  03\#TIMESTAMP#\17\#PID#\0\24
  End seq_format test.
--- 205,210 ----
  09\#TIMESTAMP#\13\#PID#\0\#TOKENSEQ23#\0\#STRMSEQNO#\3\
  09\#TIMESTAMP#\16\#PID#\0\#TOKENSEQ23#\0\#STRMSEQNO#\3\
  03\#TIMESTAMP#\14\#PID#\0\24
  03\#TIMESTAMP#\17\#PID#\0\24
+ 03\#TIMESTAMP#\14\#PID#\0\24
  End seq_format test.

Journal record lines starting with 03 correspond to EOF.

There are 3 such records corresponding to the 3 lost transaction files a.lost, b.lost and mumps.lost.
All 3 EOF records have the same journal seqno (last field) of 24. But they have different database
transaction numbers. mumps.lost has it as 17 whereas a.lost and b.lost has it as 14. In the below
test run, the EOF record of mumps.lost showed up before the EOF record of a.lost or b.lost.
Since the sort was done only on the 5th field (the journal seqno), sort could arrange lines with
matching 5th field arbitrarily and the test should take that into account to create a determinstic
output. The fix therefore is to filter the EOF records from the first output and sort just them
separately in a second round based on the 2nd field.