[V63005] Fix basic/basic subtest failure on 32-bit builds (ARMV7L/ARMV6L)
After merging GT.M V6.3-005, the basic test started failing on the ARM platform
which is a 32-bit build of YottaDB. It failed only when gtm_test_dynamic_literals
was set to DYNAMIC_LITERALS. It passed when it was set to NODYNAMIC_LITERALS.
The diff was the following.

744c744
< 150372762,SUB4+4^ztiref,%YDB-E-ERRWZTRAP, Error while processing $ZTRAP,-YDB-E-STACKCRIT, Stack space critical
---
> 150373738,SUB4+4^ztiref,%YDB-E-STACKCRIT, Stack space critical

Interestingly 64-bit builds (on x86_64) of YottaDB did not fail this test.
The reference file for this particular section (SUB4+4^ztiref) already had different outputs
on x86_64 for DYNAMIC_LITERALS and NODYNAMIC_LITERALS. Since this was related to gtm_ztrap_new
env var being set to "adaptive", an uncommon usage, this was not further investigated even in
the prior commit when the x86_64 reference file was created. For the same reasons, it is not
investigated even now and the reference file is adjusted to have different outputs on ARM
for the DYNAMIC_LITERALS and NODYNAMIC_LITERALS cases. One thing of interest is that the
32-bit Linux output is exactly the opposite of 64-bit Linux. Again, not sure why and not
considering it important to find out.