Modified outref file to account for timing issues with expect script (#333)
We had a test failure where the line "YDB>%YDB-I-CTRLC, CTRL_C encountered" in the reference file
showed up without the "YDB>" prefix. We did not investigate why that happened since the presence
of the "CTRL_C encountered" is key to the test so the faliure is fixed by allowing for an arbitrary prefix
(including "YDB>" nothing at all).