Fix v63002/gtm8644 subtest failure due to occasional extra pipe symbol in ps --forest output (#330)
In weekend testing, this subtest failed due to the presence of an extra "|" in the ps --forest output. 
That was because of a different caller script that has other children processes which are effectively
siblings of the gtm8644.csh subtest process and that caused the extra "|" which in turn affected the
fields the awk command printed resulting in a different output and a test failure. To avoid failures
due to such callers, all "|" characters in the ps --forest output are removed before going to awk.