Fix occasional gtcm_omi/basic test failure in case backward processing restarts
This test used to fail occasionally before with a JNLUNXPCTERR error because the journal file changed
(due to concurrent updates) while the MUPIP JOURNAL -EXTRACT was running. That was fixed as part of
https://github.com/YottaDB/YottaDB/issues/284. After that, in case a file state change is recognized
by the MUPIP JOURNAL -EXTRACT, it restarts backward processing. But if it does that, it prints 2 extra
lines (the line containing "Restarting Backward processing" and the immediately following line which
says "Backward processing started"). Below is a sample output from an actual test failure.

 ##FILTERED##%YDB-I-MUJNLSTAT, Initial processing started at ... ... .. ..:..:.. 20..
 ##FILTERED##%YDB-I-MUJNLSTAT, Backward processing started at ... ... .. ..:..:.. 20..
 %YDB-I-MUJNLSTAT, Restarting Backward processing at Sat Jul 21 09:52:16 2018
 ##FILTERED##%YDB-I-MUJNLSTAT, Backward processing started at ... ... .. ..:..:.. 20..
 ##FILTERED##%YDB-I-MUJNLSTAT, Forward processing started at ... ... .. ..:..:.. 20..
 %YDB-I-FILECREATE, Journal extract file exta2.out created
 %YDB-S-JNLSUCCESS, Extract successful
 %YDB-S-JNLSUCCESS, Verify successful
 ##FILTERED##%YDB-I-MUJNLSTAT, End processing at ... ... .. ..:..:.. 20..

And since the output is part of the reference file, the test fails.
The test does a lot of journal extracts and I suspect any of them could have this extra-2-lines issue.
So this test is fixed by filtering out all lines containing "Backward processing" in them that way the
output is stable whether or not the backward processing gets restarted. The only side effect of that is
that the line containing "Backward processing" which occurs after the line saying "Initial processing"
would also get filtered out. But that is considered okay since the important thing is the JNLSUCCESS at the end.