New r124/ydb312_gtm8182a subtest (one of many tests of YottaDB/YottaDB#312) (#322)
This excerpt from the initial GTM-8182 release notes was found to be not entirely accurate.

"When Instance Freeze is enabled (gtm_custom_errors is appropriately defined), a
process attaches a region to an instance at the first access to the region; the access
may be a read or a VIEW/$VIEW(). Otherwise, the process attaches to a region at the
first update to that region. "

In this test, we implement a scenario where an access (no update) of a database file causes an attach to the corresponding journal pool without the #312 code fixes. With the #312 fixes, only an update causes the attach to the corresponding journal pool.