New r124/ydb312_gtm8182b subtest (one of many tests of YottaDB/YottaDB#312) (#323)
The below issue was noticed in a code review of the GTM-8182 changes in V6.3-003.

There is a memory leak in jnlpool_init() where we allocate a tmp_jnlpool and associated region
structures but do not add it to any list in case the jnlpool is not up. And this is clearly
illustrated by a PEEKBYNAME (i.e. $zpeek) call.