New r124/ydb312_gtm8182e subtest (one of many tests of YottaDB/YottaDB#312) (#328)
Test a scenario that attempts to access a global from a global directory with no instance file mapping and when both ydb_repl_instance and gtm_repl_instance env vars are unset. This produces the below GTMASSERT2 fatal error in V6.3-003 which introduced the GTM-8182 enhancement.

%GTM-F-GTMASSERT2, GT.M V6.3-003 Linux x86_64 - Assert failed /Distrib/YottaDB/V63003/sr_port/gvcst_init.c line 1249 for expression (replpool_valid || REPL_INST_AVAILABLE(addr))

This is fixed by #312 which issues a non-fatal REPLINSTUNDEF error instead of this GTMASSERT2.