Fix occasional v63003/gtm8805 subtest hang if test framework randomly chooses spanning regions (#351)
If the test framework randomly chooses spanning regions (i.e. gtm_test_spanreg env var is set to 1)
and if a pre-computed gde mapping file is available for the random choice of chset and collation
that would be used by dbcreate. This means it is possible ^a(1) gets mapped to AREG while ^a(2)
gets mapped to BREG while ^a(3) gets mapped to DEFAULT. This would cause issues with this test
since it uses DSE to get the crit lock on DEFAULT and so any updates going to DEFAULT will never
happen which means the test would never progress beyond a particular stage and hang eternally.
This is fixed by disabling spanning regions for this test.

Additionally, it was noticed that while the test was hung, mumps processes were spinning. This is
fixed too by changing all for loops to hang a little bit (i.e. sleep-loop instead of spin-loop).