Fix weird v63002/gtm8165 hang on armv6l (pi zero)
A particular run of the v63002/gtm8165 hung on an armv6l box in internal testing.
The mumps process in the test was hung in line 170 below.

v63002/inref/gtm8165.m
----------------------
    158 writeslashtls
      .
    170         . if $ZTRNLNM("timeout")  write /tls("client",.999)

But when I ran gdb on the hung process, I got a weird C stack trace.

(gdb) where

There was no iosocket_iocontrol() in the C-stack like one would expect if a WRITE /TLS command was running.

This is the client side in the socket connection. At the same time, the server side of the socket connection
 was busy waiting for the client to reach a specific stage.

v63002/inref/gtm8165.m
----------------------
    183 tlschild
      .
    192         for  quit:^stop

Note that line 192 is a spin-loop. Once I changed the spin loop to be a sleep loop, the issue disappears.
In that, the failure happens reliably with replay with a spin loop and does not happen even once with a sleep loop.

I am not sure why the solution fixes the hang but I suspect that when the client does the WRITE /TLS command,
it waits for some acknowledgement from the server side of the socket but because the server side is caught up
in a spin loop (and the armv6l/pi-zero is a 1-cpu system), it does not ever get to acknowledging the client's
request in a timely fashion resulting in the client side tls code exercising some buggy codepath in the
openssl/tls libraries installed on the Raspbian system.

While at this, I also cleaned up the test a bit so log files are different for different stages of the
test (instead of overwriting the same log file for all stages).