Fix v63002/gtm8165 test failure due to hardcoded portno 3000 already in use
Use portno_acquire.csh and portno_release.csh to acquire a dynamic port number and use that instead of 3000.
We had a Beaglebone Black box where port 3000 was being used as a listening port by a system service and
that caused this test to fail since this tries to use 3000 for a listening socket and that failed with
an EADDRINUSE (Address already in use) error.