New r124/ydb312_gtm8182c subtest (one of many tests of YottaDB/YottaDB#312) (#324)
Tests the fix for the following scenario.

* Have 2 instances INSTA and INSTB.
* Do not start the source server for INSTB but do start for INSTA.
* Have the custom errors scheme (gtm_custom_errors env var) and the instance-freeze-on-error (-inst_freeze_on_error flag in GDE) enabled.
* And add KEY2BIG as an error that triggers instance freeze in $ydb_dist/custom_errors_sample.txt
* Then trigger a KEY2BIG error while referencing a database in INSTB.

In GT.M V6.3-003 (which has the GTM-8182 changes) the instance freeze is incorrectly applied to INSTA.
This is because there was no jnlpool found for INSTB and the process had opened a jnlpool for INSTA. 

With YottaDB/YottaDB#312 fixes, this works correctly by leaving INSTA untouched (i.e. no incorrect freeze).