New v63005/gtm8980 subtest (tests GTM-8980 in V63005) (#345)
Tests the following release note:

VIEW and $VIEW() with a empty string or inappropriate region-list works appropriately; in V6.3-004 these could cause inappropriate results, including a segmentation violation (SIG-11). $VIEW("statshare") returns a 0 when the process has sharing disabled, a 1 when it has sharing enabled and a 2 when sharing is selectively enabled. In V6.3-004, it did not differentiate between the all and selective cases and returned 1 when sharing was disabled and selective disabling was also specified. $VIEW("statshare","") works appropriately even if the region had been selectively disabled when full sharing is disabled and the region had not been opened. In V6.3-004, this set of conditions produced a segmentation violation (SIG-11). The error messages when invalid parameters are passed to VIEW/$VIEW() print the name of the parameter; previously such error messages did not have the name of the parameter. (GTM-8980)